# docs/reference/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Index","operationId":"index__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health/ready":{"get":{"summary":"Ready","description":"Readiness probe endpoint.","operationId":"ready_health_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/hub-names":{"get":{"summary":"Hub Names","description":"Endpoint to query hub names.","operationId":"hub_names_hub_names_get","parameters":[{"name":"from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Start date in YYYY-MM-DDTHH:MMZ format","title":"From"},"description":"Start date in YYYY-MM-DDTHH:MMZ format"},{"name":"to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"End date in YYYY-MM-DDTHH:MMZ format","title":"To"},"description":"End date in YYYY-MM-DDTHH:MMZ format"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/component-names":{"get":{"summary":"Component Names","description":"Endpoint to serve component names.","operationId":"component_names_component_names_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/total-costs":{"get":{"summary":"Total Costs","description":"Endpoint to query total costs.","operationId":"total_costs_total_costs_get","parameters":[{"name":"from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Start date in YYYY-MM-DDTHH:MMZ format","title":"From"},"description":"Start date in YYYY-MM-DDTHH:MMZ format"},{"name":"to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"End date in YYYY-MM-DDTHH:MMZ format","title":"To"},"description":"End date in YYYY-MM-DDTHH:MMZ format"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user-groups":{"get":{"summary":"User Groups","description":"Endpoint to serve user group memberships.","operationId":"user_groups_user_groups_get","parameters":[{"name":"from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Start date in YYYY-MM-DDTHH:MMZ format","title":"From"},"description":"Start date in YYYY-MM-DDTHH:MMZ format"},{"name":"to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"End date in YYYY-MM-DDTHH:MMZ format","title":"To"},"description":"End date in YYYY-MM-DDTHH:MMZ format"},{"name":"hub","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Name of the hub to filter results","title":"Hub"},"description":"Name of the hub to filter results"},{"name":"username","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Name of the user to filter results","title":"Username"},"description":"Name of the user to filter results"},{"name":"usergroup","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Name of the group to filter results","title":"Usergroup"},"description":"Name of the group to filter results"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users-with-multiple-groups":{"get":{"summary":"Users With Multiple Groups","description":"Endpoint to serve users with multiple groups.","operationId":"users_with_multiple_groups_users_with_multiple_groups_get","parameters":[{"name":"hub_name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Name of the hub to filter results","title":"Hub Name"},"description":"Name of the hub to filter results"},{"name":"user_name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Name of the user to filter results","title":"User Name"},"description":"Name of the user to filter results"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users-with-no-groups":{"get":{"summary":"Users With No Groups","description":"Endpoint to serve users with no groups.","operationId":"users_with_no_groups_users_with_no_groups_get","parameters":[{"name":"hub_name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Name of the hub to filter results","title":"Hub Name"},"description":"Name of the hub to filter results"},{"name":"user_name","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Name of the user to filter results","title":"User Name"},"description":"Name of the user to filter results"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/total-costs-per-hub":{"get":{"summary":"Total Costs Per Hub","description":"Endpoint to query total costs per hub.","operationId":"total_costs_per_hub_total_costs_per_hub_get","parameters":[{"name":"from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Start date in YYYY-MM-DDTHH:MMZ format","title":"From"},"description":"Start date in YYYY-MM-DDTHH:MMZ format"},{"name":"to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"End date in YYYY-MM-DDTHH:MMZ format","title":"To"},"description":"End date in YYYY-MM-DDTHH:MMZ format"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/total-costs-per-component":{"get":{"summary":"Total Costs Per Component","description":"Endpoint to query total costs per component.","operationId":"total_costs_per_component_total_costs_per_component_get","parameters":[{"name":"from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Start date in YYYY-MM-DDTHH:MMZ format","title":"From"},"description":"Start date in YYYY-MM-DDTHH:MMZ format"},{"name":"to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"End date in YYYY-MM-DDTHH:MMZ format","title":"To"},"description":"End date in YYYY-MM-DDTHH:MMZ format"},{"name":"hub","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Name of the hub to filter results","title":"Hub"},"description":"Name of the hub to filter results"},{"name":"component","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Name of the component to filter results","title":"Component"},"description":"Name of the component to filter results"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/total-costs-per-group":{"get":{"summary":"Total Costs Per Group","description":"Endpoint to query total costs per user group.","operationId":"total_costs_per_group_total_costs_per_group_get","parameters":[{"name":"from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Start date in YYYY-MM-DDTHH:MMZ format","title":"From"},"description":"Start date in YYYY-MM-DDTHH:MMZ format"},{"name":"to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"End date in YYYY-MM-DDTHH:MMZ format","title":"To"},"description":"End date in YYYY-MM-DDTHH:MMZ format"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/costs-per-user":{"get":{"summary":"Costs Per User","description":"Endpoint to query costs per user by combining AWS costs with Prometheus usage data.\n\nThis endpoint calculates individual user costs by:\n1. Getting total AWS costs per component (compute, storage) from Cost Explorer\n2. Getting usage fractions per user from Prometheus metrics\n3. Multiplying total costs by each user's usage fraction\n\nQuery Parameters:\n    from (str): Start date in YYYY-MM-DD format (defaults to 30 days ago)\n    to (str): End date in YYYY-MM-DD format (defaults to current date)\n    hub (str, optional): Filter to specific hub namespace\n    component (str, optional): Filter to specific component (compute, home storage)\n    user (str, optional): Filter to specific user\n    usergroup (str, optional): Filter to specific user group\n    limit (int, optional): Limit number of results to top N users by total cost.\n\nReturns:\n    List of dicts with keys: date, hub, component, user, value (cost in USD)\n    Results are sorted by date, hub, component, then value (highest cost first)","operationId":"costs_per_user_costs_per_user_get","parameters":[{"name":"from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Start date in YYYY-MM-DDTHH:MMZ format","title":"From"},"description":"Start date in YYYY-MM-DDTHH:MMZ format"},{"name":"to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"End date in YYYY-MM-DDTHH:MMZ format","title":"To"},"description":"End date in YYYY-MM-DDTHH:MMZ format"},{"name":"hub","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Name of the hub to filter results","title":"Hub"},"description":"Name of the hub to filter results"},{"name":"component","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Name of the component to filter results","title":"Component"},"description":"Name of the component to filter results"},{"name":"user","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Name of the user to filter results","title":"User"},"description":"Name of the user to filter results"},{"name":"usergroup","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Name of user group to filter results","title":"Usergroup"},"description":"Name of user group to filter results"},{"name":"limit","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Limit number of results to top N users by total cost.","title":"Limit"},"description":"Limit number of results to top N users by total cost."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/total-usage":{"get":{"summary":"Total Usage","description":"Endpoint to query total usage.\nExpects 'from' and 'to' query parameters in the api_provider YYYY-MM-DD.\nOptionally accepts 'hub', 'component' and 'user', query parameters.","operationId":"total_usage_total_usage_get","parameters":[{"name":"from","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Start date in YYYY-MM-DDTHH:MMZ format","title":"From"},"description":"Start date in YYYY-MM-DDTHH:MMZ format"},{"name":"to","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"End date in YYYY-MM-DDTHH:MMZ format","title":"To"},"description":"End date in YYYY-MM-DDTHH:MMZ format"},{"name":"hub","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Name of the hub to filter results","title":"Hub"},"description":"Name of the hub to filter results"},{"name":"component","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Name of the component to filter results","title":"Component"},"description":"Name of the component to filter results"},{"name":"user","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Name of the user to filter results","title":"User"},"description":"Name of the user to filter results"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}